Fix perfdata parser not recognize scientific notation
The scientific notation is basically allowed in our performance data
parser. In an edge case where scientific notation is delivered with
an big E letter our parser does not recognize it and drops it as
false performance data.

This wrong interpretation as false performance data affects all parts of
the performance data, the value itself, the warning value, the critical
value, the minimum value and the maximum value.

Since all parts of the performance data are extracted from the valueStr,
it is sufficient to transform it to the lower case.